refactor(desktop): group OS/external integrations into main/integrations
