+++ apps/desktop/src/main/index.ts
@@ -9,8 +9,8 @@ import { initializeAutoUpdater } from './app/updater';
 import { logger, flushLogs } from './app/logger';
 import { createTray, destroyTray } from './app/tray';
 import { initializeSystemBehavior, attachTrayWindowBehavior } from './app/system-behavior';
-import { initializeMediaControls, cleanupMediaControls } from './media-controls';
-import { initializeDiscordRpc, cleanupDiscordRpc } from './discord-rpc';
+import { initializeMediaControls, cleanupMediaControls } from './integrations/media-controls';
+import { initializeDiscordRpc, cleanupDiscordRpc } from './integrations/discord-rpc';
 import { registerAudioProtocol } from './protocols/audio-protocol';
 import { registerRadioProtocol } from './protocols/radio-protocol';
 import { registerArtProtocol, pruneOrphanedAlbumArt } from './protocols/art-protocol';
