+++ apps/desktop/src/main/integrations/discord-presence-builder.test.ts
