+++ apps/desktop/src/main/integrations/discord-presence-builder.ts
